fix: Add version compatibility checks for LiteLLM and OpenAI
This adds runtime checks in `LiteLlm.__init__` to make it so that the installed versions of `litellm` have to be less than 2.0.0 and `openai` are less than or equal to 1.99.9. Inn `pyproject.toml` added more comments and add version bounding in the optional extension , and `packaging` has been added as a dependency for version parsing.

PiperOrigin-RevId: 796554298