fix: Convert examples for A2A agent card
The AgentSkill in an A2A AgentCard expects examples to be a list of
queries as strings. Therefore, agent examples, e.g., as provided by an
ExampleTool, must be converted.

This change performs that extraction of just the inputs and converting
them to a string to add to the AgentSkill.